Throw error when subsetting relative risk data to empty dataframe (#298)
    Category: feature/test
    JIRA issue: MIC-3468

Changes and notes

Move validation out of rebinning step to after loading relative risk data. Throw an error if the data in this validation step is empty. Add a test that checks if error is thrown when relative risk data is missing data for the specified target. Change "cause" column in mock artifact to "affected_entity".

Testing
Check that test passed. Check that other ways of ending up with an empty dataframe will either not occur or throw a different error. Check that changing mock artifact won't affect existing tests.