fix validate_resource_bin bug (#301)
    Category: bugfix
    JIRA issue: MIC-3795

Changes and notes
Previously added argument to validate_rebin_source which was only relevant when data is for relative risk. A call to this function was not updated appropriately when validating exposure data. Keep the original function in the case of exposure validation and create new function which performs relative risk-specific validation and then calls the original function.

Testing
Successfully ran simulation setup for US CVD which was failing because of this bug.